Add `keymap_nathan.c`
A Planck layout for the 'grid' bottom row.

Normal layer is for alphas with 'tab', 'ctrl', 'shift', 'meta' as the
first column. Arrow key cluster is on the bottom right. The four braces
have the left variant on a key each: [ ( {<

Left/Lower layer has VisualStudio and Windows shortcuts, numpad layout
numbers and a row of macros which run python a `keypress.py` python
script with a number argument (dependant on the key). There are
navigation keys on the arrow keys.

Right/Raise layer has F1-F12 and all symbols. There are media controls
on the arrow keys. Each of the braces has the right variant on a key: ]
) } >